datasets/preprocess/inpainting/VideoPainter/requirements.txt
torch==2.4.0
torchvision>=0.19.0
transformers==4.42.2
accelerate>=0.34.2
git+https://github.com/microsoft/DeepSpeed.git
git+https://github.com/facebookresearch/pytorchvideo.git


opencv-python
pillow==9.5.0
imageio>=2.35.1
imageio-ffmpeg==0.5.1
ffmpeg-python
moviepy>=1.0.3
decord==0.6.0
kornia==0.7.3
scikit-image

git+https://github.com/openai/CLIP.git
openai>=1.45.0
peft
xformers
bitsandbytes
sentencepiece>=0.2.0

numpy==1.26.0
scipy
einops
rotary_embedding_torch
pytorch_lightning
easydict
pandas
matplotlib
gradio==3.40.0
wandb
nvitop
omegaconf==2.3.0
beartype==0.18.5
fsspec==2024.5.0
safetensors==0.4.3
huggingface_hub==0.24.1
av==13.1.0
gdown